refactor: phase 4 - extract storage service
